feat(landing): move Copy Prompt into nav bar, keep Get Started as hero CTA
- replace the desktop nav 'Get Started for free' CTA with the Copy Prompt button
  (copy logic + single-flight feedback now live in SiteNav)
- remove the standalone hero Copy Prompt block + prompt preview
- hero keeps 'Get Started for free' + 'Contact sales'; mobile menu unchanged
